refactor(editor): Extract tags code into features/shared (no-changelog)
