+++ packages/frontend/editor-ui/src/features/ai/evaluation.ee/evaluation.store.test.ts
@@ -1,7 +1,7 @@
 import { createPinia, setActivePinia } from 'pinia';
 import { useEvaluationStore } from './evaluation.store'; // Adjust the import path as necessary
 import { useRootStore } from '@n8n/stores/useRootStore';
-import { useAnnotationTagsStore } from '@/stores/tags.store';
+import { useAnnotationTagsStore } from '@/features/shared/tags/tags.store';
 import type { TestRunRecord } from './evaluation.api';
 import { mockedStore } from '@/__tests__/utils';
