+++ packages/frontend/editor-ui/src/features/shared/tags/components/TagsManager/TagsView/TagsView.vue
@@ -1,10 +1,10 @@
 <script setup lang="ts">
 import { ref, computed } from 'vue';
 import type { ITag } from '@n8n/rest-api-client/api/tags';
-import type { ITagRow } from '@/Interface';
+import type { ITagRow } from '../../../tags.types';
 import { useI18n } from '@n8n/i18n';
-import TagsTableHeader from '@/components/TagsManager/TagsView/TagsTableHeader.vue';
-import TagsTable from '@/components/TagsManager/TagsView/TagsTable.vue';
+import TagsTableHeader from './TagsTableHeader.vue';
+import TagsTable from './TagsTable.vue';
 import { useRBACStore } from '@/stores/rbac.store';
 import type { BaseTextKey } from '@n8n/i18n';
