+++ packages/frontend/editor-ui/src/features/shared/tags/tags.constants.ts
@@ -1,0 +1,4 @@
+export const MAX_TAG_NAME_LENGTH = 24;
+
+export const TAGS_MANAGER_MODAL_KEY = 'tagsManager';
+export const ANNOTATION_TAGS_MANAGER_MODAL_KEY = 'annotationTagsManager';
