+++ packages/frontend/editor-ui/src/features/shared/tags/tags.types.ts
@@ -1,0 +1,11 @@
+import type { ITag } from '@n8n/rest-api-client/api/tags';
+
+export interface ITagRow {
+	tag?: ITag;
+	usage?: string;
+	create?: boolean;
+	disable?: boolean;
+	update?: boolean;
+	delete?: boolean;
+	canDelete?: boolean;
+}
